requirements.txt
fastapi
uvicorn
openai
python-dotenv
streamlit
requests
chromadb
PyPDF2
tavily-python
pydantic